Created GeometricBrownianBridge class, and implemented simulation of wiener process using the brownian bridge method

+++ src/main/java/org/example/StochasticProcess.java
@@ -19,7 +19,7 @@ public abstract class StochasticProcess {
 
     public abstract double simulateStep(double dt, double Z);
 
-    public List<Double> simulateSteps(double count, double dt, double[] randoms) {
+    public List<Double> simulateSteps(int count, double dt, double[] randoms) {
         List<Double> list = new ArrayList<>();
         for(int i = 0; i < count; i++) {
             list.add(simulateStep(dt, randoms[i]));
